+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## UNRELEASED
+- Adds a new `wordList` option. 
+
 ## 2.0.1 (2024-01-25)
 
 - Fixed return type of `generate` so that it is consistent with the possibility of returning either `string` or `string[]`